Solved part 1 and cheated part 2
Part 1 was fine, part 2 was not. In the end I had to use the fact
they're equations and get pulp to solve it, which i learnt from the
reddit answers. Of course, I first spent 3+ hours trying to solve it
myself, even getting chatGPT to help me understand better approaches and
ways of tackling the problem but that was all for naught and eventually
I gave up and used pulp. Problem is solved but I need to review it some
more ideally. Yay.